chore: move compile support to standalone mod
